LayoutTests/svg/animations/svglength-animation-px-to-in-expected.txt

SVG 1.1 dynamic animation tests

Test SVGLength animation from px to inch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 170
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 240
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
